+++ src/View/Helper/PaginatorHelper.php
@@ -16,7 +16,6 @@ namespace MeTools\View\Helper;
 use Cake\Utility\Hash;
 use Cake\View\Helper\PaginatorHelper as CakePaginatorHelper;
 use Cake\View\View;
-use MeTools\View\OptionsParser;
 
 /**
  * Provides functionalities to the generation of pagers
@@ -68,7 +67,7 @@ class PaginatorHelper extends CakePaginatorHelper
      */
     public function next($title = 'Next >>', array $options = [])
     {
-        $options = new OptionsParser($options, ['escape' => false, 'icon-align' => 'right']);
+        $options = optionsParser($options, ['escape' => false, 'icon-align' => 'right']);
         list($title, $options) = $this->Html->addIconToText($title, $options);
 
         return parent::next($title, $options->toArray());
@@ -82,7 +81,7 @@ class PaginatorHelper extends CakePaginatorHelper
      */
     public function prev($title = '<< Previous', array $options = [])
     {
-        $options = new OptionsParser($options, ['escape' => false]);
+        $options = optionsParser($options, ['escape' => false]);
         list($title, $options) = $this->Html->addIconToText($title, $options);
 
         return parent::prev($title, $options->toArray());